Add Pointer Caching and Loop Unrolling to Activation Functions

+++ c_reference/include/quantized_datatypes.h
@@ -9,9 +9,9 @@
 // Macro for input type.
 typedef int16_t INT_T;
 // Macro for unsigned iterator type.
-typedef uint16_t ITER_T;
+typedef uint32_t ITER_T;
 // Macro for signed iterator type.
-typedef int16_t S_ITER_T;
+typedef int32_t S_ITER_T;
 // Macro for intermediate buffer type.
 typedef int32_t INTM_T;
 // Macros for scale variable type.
